Добавляет тоасты, изменяет авторизацию
