+++ project/package.json
@@ -15,6 +15,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-spinners": "^0.11.0",
+    "react-toastify": "^8.1.0",
     "reselect": "^4.0.0"
   },
   "scripts": {
