--- project/src/components/login-failed/login-failed.module.css
@@ -1,27 +1,0 @@
-.login-failed {
-  display: flex;
-  flex-grow: 1;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  border: 2px solid red;
-  border-radius: 5px;
-  margin-top: 24px;
-}
-
-.login-failed--info {
-  display: block;
-  padding: 20px;
-  margin-bottom: 20px;
-  line-height: 1.1875;
-  text-align: center;
-}
-
-.login-failed--info-title {
-  font-size: 22px;
-  color: black;
-}
-
-.login-failed--info-text {
-  font-size: 14px;
-}
