+++ project/src/index.tsx
@@ -13,6 +13,9 @@ import { requireAutorization } from './store/action';
 import { AuthorizationStatus } from './const';
 import { redirect } from './store/middlewares/redirect';
 
+import {ToastContainer} from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
 const api = createAPI(
   () => store.dispatch(requireAutorization(AuthorizationStatus.NoAuth)),
 );
@@ -31,6 +34,7 @@ const store = createStore(
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
+      <ToastContainer/>
       <App/>
     </Provider>
   </React.StrictMode>,
